needs a vertex structure
Foun out why it was only showing half the polygon's vertices: I had
forgotten that the vertex attributes were for a single vertex.  Whoops.
I was specifying the polygon face with two positions and a normal, but
the vertex attributes were telling OpenGL that each vertex had two vec4
positions and 1 vec4 normal, and it was fine with that.  My bad.  Now
adding MyVertex back into the program.