got the particles moving, now making them move correctly
The problem was that I wasn't telling the compute shader how many
particles there were, so uParticleCount was 0 and the compute shader
happily did nothing (hey, it's valid code).  I removed the "if (index <
uParticleCount)" and the hard-coded particle position updates worked, so
I put that line back in and uploaded the particle count with
glUniform1ui(...) and it worked.  Now I need to figure out how to give
all the particles random velocities.  I may have have to use the
particle emitters on the CPU side to give them initial values.