IT'S WORKING!
After some interesting stuff with math and linear interpolation
revolving around how to get the particles to reset in a circular region
around the point emitter's reset position, I finally got it.  Also some
cleanup.  More cleanup is still needed and I need to implement
transformational matrices in both the compute and render shaders, plus I
need to account for the bar emitter and possibly multiples of both point
and bar emitters, but the basics are working now.  Wewt!