atomic counter active + alpha blending
Slightly altered compute shader to take into account whether particles
were "active" or not.  If not, then they are reactivating and an atomic
counter is incremented (courtesy of geeks3D example code).  The atomic
counter now limits the number of particles reactivated per frame.  Also
altered the geometry and particle vert shaders and frag shaders so that
inactive particles draw black AND don't make the polygon fragment that
they "ran into" draw black as well.  I don't understand why the alpha
blending works the way it does, but it works for now.  It needs cleanup
and proper explanation.