figured out how to get multiple emitters working
The trouble was twofold: (1) Multiple calls to the compute shader were
causing previously updated particles to update again, which required the
compute shader's purpose to be divided into resetting particles and
updating their positions.  (2) There was trouble getting the compute
shader to play ball with a couple new uniforms.  It kept culling the
functionality after "if (uOnlyResetParticles == 1)", and then it didn't.
I don't know what I did and why it started working, but it did.