Now pre-computing geometry transformation
This pre-computation follows in the footsteps of the "render particles
2D advanced CPU" project, which found a considerable performance boost
when it was computing vector transformations once per frame instead of
once per particle per frame.  The GPU is already much faster than the
CPU on these particle problems and I haven't taxed it to under ~60fps
yet, so I saw no performance boost, but I'm confident that there is an
efficiency boost.