fixed it
The problem was that layout binding is global across shaders.  I didn't
know that.  I had one atomic counter for the number of reset particles
in the "particle reset" compute shader and another atomic counter for
the number of active particles in "particle updater", and they used the
same layout binding (zero).  The ComputeUpdaterParticleReset object
initialized before ComputeUpdaterParticleUpdater, so the buffer base
binding for the latter's atomic counters ran over the binding for the
former's, leaving the reset binding for the reset atomic counter stuck
using the "number active particles" atomic counter from the particle
updating shader.  Whoops.  Lesson learned.